CXP-758 Moved ConnectedAppAuthorizations tests from ConnectedAppSettings
